+++ src/service.rs
@@ -31,16 +31,23 @@ use http_types::headers::HeaderValue;
 use manta_accounting::{
     asset::Asset,
     fs::{cocoon::File, File as _, SaveError},
+    key::AccountTable,
     transfer::canonical::TransferShape,
+    wallet::signer::functions::default_authorization_context,
 };
 use manta_pay::{
-    config::Transaction,
+    config::{Config as MantaPayConfig, Transaction},
     key::{Mnemonic, TestnetKeySecret},
     signer::{
         base::{Signer, SignerParameters, SignerState, UtxoAccumulator},
         client::network::{Message, Network, NetworkSpecific},
+        AssetMetadata, TokenType,
     },
 };
+
+/// previous version of manta-pay state. Should use versioning system for SignerState
+use previous_state_manta_pay::signer::base::SignerState as OldSignerState;
+
 use manta_util::{from_variant, serde::Serialize};
 use parking_lot::Mutex;
 use std::{
@@ -65,7 +72,6 @@ pub use manta_pay::{
         SyncError, SyncResponse, TransactionDataResponse,
     },
 };
-use manta_pay::signer::{AssetMetadata, TokenType};
 
 /// Synchronization Request
 pub type SyncRequest = Message<signer::SyncRequest>;
@@ -178,26 +184,49 @@ pub fn display_transaction(
         TokenType::NFT => 1
     };
     match transaction {
-        Transaction::ToPrivate(Asset { value, .. }) => format!(
-            "Privatize {} on {} network",
-            metadata.display(*value, decimal),
-            network
-        ),
+        Transaction::ToPrivate(Asset { value, .. }) => match metadata.token_type {
+            TokenType::FT(decimals) => {
+                format!(
+                    "Privatize {} on {} network",
+                    metadata.display(*value, decimals),
+                    network
+                )
+            }
+            TokenType::NFT => {
+                format!("Privatize NFT on {} network", network)
+            }
+        },
         Transaction::PrivateTransfer(Asset { value, .. }, receiving_key) => {
-            format!(
-                "Send {} to {} on {} network",
-                metadata.display(*value, decimal),
-                address_to_base58(receiving_key),
-                network
-            )
+            match metadata.token_type {
+                TokenType::FT(decimals) => {
+                    format!(
+                        "Send {} to {} on {} network",
+                        metadata.display(*value, decimals),
+                        address_to_base58(receiving_key),
+                        network
+                    )
+                }
+                TokenType::NFT => {
+                    format!(
+                        "Send NFT to {} on {} network",
+                        address_to_base58(receiving_key),
+                        network
+                    )
+                }
+            }
         }
-        Transaction::ToPublic(Asset { value, .. }) => {
-            format!(
-                "Public {} on {} network",
-                metadata.display(*value, decimal),
-                network
-            )
-        }
+        Transaction::ToPublic(Asset { value, .. }, _) => match metadata.token_type {
+            TokenType::FT(decimals) => {
+                format!(
+                    "Public {} on {} network",
+                    metadata.display(*value, decimals),
+                    network
+                )
+            }
+            TokenType::NFT => {
+                format!("Public NFT on {} network", network)
+            }
+        },
     }
 }
 
@@ -232,9 +261,11 @@ where
                     .is_ok()
                 {
                     self.authorizer.sleep().await;
+                    println!("Password is ok");
                     return Ok(());
                 }
             } else {
+                println!("Password is now known, returning auth error");
                 return Err(Error::AuthorizationError);
             }
             delay_password_retry().await;
@@ -302,6 +333,7 @@ where
                         &config.data_path.dolphin,
                         &password_hash,
                         mnemonic.clone(),
+                        &parameters,
                     )
                     .await?;
 
@@ -310,6 +342,7 @@ where
                         &config.data_path.calamari,
                         &password_hash,
                         mnemonic.clone(),
+                        &parameters,
                     )
                     .await?;
 
@@ -318,6 +351,7 @@ where
                         &config.data_path.manta,
                         &password_hash,
                         mnemonic.clone(),
+                        &parameters,
                     )
                     .await?;
                     break (
@@ -351,6 +385,7 @@ where
                         &config.data_path.dolphin,
                         &password_hash,
                         recovery_mnemonic.clone(),
+                        &parameters,
                     )
                     .await?;
 
@@ -359,6 +394,7 @@ where
                         &config.data_path.calamari,
                         &password_hash,
                         recovery_mnemonic.clone(),
+                        &parameters,
                     )
                     .await?;
 
@@ -367,6 +403,7 @@ where
                         &config.data_path.manta,
                         &password_hash,
                         recovery_mnemonic.clone(),
+                        &parameters,
                     )
                     .await?;
 
@@ -422,7 +459,7 @@ where
         };
         let existing_signer = Signer::from_parts(
             parameters.clone(),
-            Self::load_state(existing_state_path, password_hash)
+            Self::load_state(existing_state_path, password_hash, &parameters)
                 .await
                 .expect("Unable to get dolphin state")?,
         );
@@ -430,6 +467,8 @@ where
             existing_signer
                 .state()
                 .accounts()
+                .as_ref()
+                .expect("No accounts in signer state!")
                 .keys()
                 .expose_mnemonic()
                 .clone(),
@@ -444,6 +483,7 @@ where
         data_path: &Path,
         password_hash: &PasswordHash<Argon2>,
         recovery_mnemonic: Option<Mnemonic>,
+        parameters: &SignerParameters,
     ) -> Result<Option<SignerState>> {
         if should_recreate {
             info!("state missing! recreating state.")?;
@@ -451,12 +491,13 @@ where
                 data_path,
                 password_hash,
                 recovery_mnemonic.expect("unable to retrieve mnemonic for account recreation."),
+                parameters,
             )
             .await
             .expect("Unable to recreate signer instance from existing mnemonic.");
             Ok(Some(state))
         } else {
-            Self::load_state(data_path, password_hash).await
+            Self::load_state(data_path, password_hash, parameters).await
         }
     }
 
@@ -465,6 +506,8 @@ where
     #[inline]
     pub async fn cancel_signing(&mut self) {
         self.state.lock().currently_signing = false;
+        // Forcefully sleep because the authorizer gets stuck awake if we exit recovery window
+        self.authorizer.lock().await.authorizer.sleep();
     }
 
     /// Starts the signer server with `config` and `authorizer`.
@@ -515,16 +558,21 @@ where
         data_path: &Path,
         password_hash: &PasswordHash<Argon2>,
         mnemonic: Mnemonic,
+        parameters: &SignerParameters,
     ) -> Result<SignerState> {
         info!("creating signer state")?;
-        let state = SignerState::new(
-            TestnetKeySecret::new(mnemonic, ""),
-            UtxoAccumulator::new(
-                task::spawn_blocking(crate::parameters::load_utxo_accumulator_model)
-                    .await?
-                    .ok_or(Error::ParameterLoadingError)?,
-            ),
-        );
+        let mut state = SignerState::new(UtxoAccumulator::new(
+            task::spawn_blocking(crate::parameters::load_utxo_accumulator_model)
+                .await?
+                .ok_or(Error::ParameterLoadingError)?,
+        ));
+        let accounts = AccountTable::new(TestnetKeySecret::new(mnemonic, ""));
+        state.load_authorization_context(default_authorization_context::<MantaPayConfig>(
+            &accounts,
+            &parameters.parameters,
+        ));
+        state.load_accounts(accounts);
+
         info!("saving signer state")?;
         let data_path = data_path.to_owned();
         let password_hash_bytes = password_hash.as_bytes();
@@ -539,17 +587,51 @@ where
     async fn load_state(
         data_path: &Path,
         password_hash: &PasswordHash<Argon2>,
+        parameters: &SignerParameters,
     ) -> Result<Option<SignerState>> {
         info!("loading signer state from disk")?;
-        let data_path = data_path.to_owned();
+        let data_path_buf = data_path.to_owned();
+        let password_hash_bytes = password_hash.as_bytes();
+
+        let state_result = task::spawn_blocking(move || {
+            File::load::<_, SignerState>(&data_path_buf, &password_hash_bytes)
+        })
+        .await?;
+
+        if let Ok(correct_state) = state_result {
+            Ok(Some(correct_state))
+        } else {
+            // fallback to try from old state version
+            Self::new_state_from_old_state(data_path, password_hash, parameters).await
+        }
+    }
+
+    /// Attempts to create new signer state from the old signer state version
+    #[inline]
+    async fn new_state_from_old_state(
+        data_path: &Path,
+        password_hash: &PasswordHash<Argon2>,
+        parameters: &SignerParameters,
+    ) -> Result<Option<SignerState>> {
+        info!("loading mnemonic from old state")?;
+        let data_path_buf = data_path.to_owned();
         let password_hash_bytes = password_hash.as_bytes();
 
         if let Ok(state) = task::spawn_blocking(move || {
-            File::load::<_, SignerState>(&data_path, &password_hash_bytes)
+            File::load::<_, OldSignerState>(&data_path_buf, &password_hash_bytes)
         })
         .await?
         {
-            Ok(Some(state))
+            let mnemonic = state.accounts().keys().expose_mnemonic().clone();
+
+            let encoded: Vec<u8> = bincode::serialize(&mnemonic).expect("encoding mnenomic failed");
+            let new_mnemonic: Mnemonic =
+                bincode::deserialize(&encoded[..]).expect("decoding mnenomic failed");
+
+            let new_state =
+                Self::create_state(data_path, password_hash, new_mnemonic, parameters).await?;
+
+            Ok(Some(new_state))
         } else {
             Ok(None)
         }
@@ -712,6 +794,8 @@ where
         let stored_mnemonic = self.state.lock().signer[network]
             .state()
             .accounts()
+            .as_ref()
+            .expect("No accounts in signer")
             .keys()
             .expose_mnemonic()
             .clone();
@@ -721,7 +805,9 @@ where
     /// Runs the receiving key sampling protocol on the signer.
     #[inline]
     pub async fn address(self, request: ReceivingKeyRequest) -> Result<Address> {
-        let response = self.state.lock().signer[request.network].address();
+        let response = self.state.lock().signer[request.network]
+            .address()
+            .expect("No address present in signer!");
         info!("[RESPONSE] responding to `receivingKeys` with: {response:?}")?;
         Ok(response)
     }
@@ -730,7 +816,9 @@ where
     /// the result to base 58.
     #[inline]
     pub async fn get_address(&mut self, request: ReceivingKeyRequest) -> Result<String, ()> {
-        let response = self.state.lock().signer[request.network].address();
+        let response = self.state.lock().signer[request.network]
+            .address()
+            .ok_or(())?;
         let key = address_to_base58(&response);
         Ok(key)
     }